day 11
Not super proud of this one, part two was solved by translating someones
Elixir code to go. I really could not figure this one out. But now I
have a better understanding. This was a great lesson on go memory
managment. And my oh my, are strings expensive memory wise.